+++ packages/contracts/src/whitelist.ts
@@ -6,7 +6,8 @@ import { utils } from "ethers";
 const exec = util.promisify(_exec);
 
 const networks: string[] = [
-  "mainnet",
+  // "mainnet",
+  "moonbeam",
   "optimism",
   "bsc",
   "xdai",
@@ -17,29 +18,29 @@ const networks: string[] = [
   "arbitrum-one",
   "avalanche",
 ];
+
+const routers: string[] = [];
+
 const run = async () => {
-  const cmdArg = process.argv.slice(2);
-
-  // first argument is router address
-  const _routerAddress = cmdArg[0];
-
-  if (!_routerAddress) {
-    console.log("please add router address, checkout readme for more");
+  if (routers.length === 0) {
+    console.log("please add router addresses to whitelist");
     return;
   }
 
-  const routerAddress = utils.getAddress(_routerAddress);
+  const routerAddresses = routers.map((r) => utils.getAddress(r));
 
-  for (const n of networks) {
-    console.log("Running add router script for", n);
-    const { stdout: out, stderr: err } = await exec(`yarn hardhat add-router --network ${n} --router ${routerAddress}`);
+  networks.forEach((n) => {
+    routerAddresses.forEach(async (r) => {
+      console.log("Running add router script for", n);
+      const { stdout: out, stderr: err } = await exec(`yarn hardhat add-router --network ${n} --router ${r}`);
 
-    if (out) {
-      console.log(`stdout: ${n} ${out}`);
-    }
-    if (err) {
-      console.error(`stderr: ${n} ${err}`);
-    }
-  }
+      if (out) {
+        console.log(`stdout: ${n} ${out}`);
+      }
+      if (err) {
+        console.error(`stderr: ${n} ${err}`);
+      }
+    });
+  });
 };
 run();